requirements.txt
python-telegram-bot==20.3
gspread==5.12.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
